--- okhttp-testing-support/src/test/kotlin/okhttp3/testing/JdkMatchRuleTest.kt
@@ -1,41 +1,0 @@
-/*
- * Copyright (C) 2019 Square, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package okhttp3.testing
-
-import org.junit.Rule
-import org.junit.Test
-
-class JdkMatchRuleTest {
-  @Suppress("RedundantVisibilityModifier")
-  @JvmField
-  @Rule public var jdkMatchRule = JdkMatchRule()
-
-  @Test
-  fun testGreenCase() {
-  }
-
-  @Test
-  fun testGreenCaseFailingOnLater() {
-    jdkMatchRule.expectFailure(fromMajor(VersionInfo.majorVersion + 1))
-  }
-
-  @Test
-  fun failureCase() {
-    jdkMatchRule.expectFailure(fromMajor(VersionInfo.majorVersion))
-
-    check(false)
-  }
-}
