+++ okhttp/src/test/java/okhttp3/CallKotlinTest.kt
@@ -17,6 +17,7 @@ package okhttp3
 
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
+import okhttp3.testing.PlatformRule
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Before
 import org.junit.Rule
